build: fixes for building for (non-Android-)Linux
This is a preparatory step for adding build scripts for openmobilemaps
for (non-Android-)Linux.

Fix various missing includes and other small errors that are accepted by
the Android/iOS build toolchains, but are errors when attempting to
build with a generic GCC/Clang.
In particular, change `abs` to `std::abs` when used with float arguments
as suggested by clang (-Wabsolute-value); this avoids the implicit and
likely undesired truncation of floats with `abs`.
Change `#ifdef __ANDROID__` to `#ifdef __linux__` as OpenGL feature
toggle for now.

Fix misc small warnings, making this build cleanly (on GCC 13.2, clang
18.1) with "only" disabling "-Wno-reorder -Wno-deprecated" (note that's
with default warnings, without -Wall).

+++ android/src/main/cpp/graphics/objects/Quad2dStretchedInstancedOpenGl.h
@@ -128,15 +128,15 @@ class Quad2dStretchedInstancedOpenGl : public GraphicsObjectInterface,
     int instStretchYsHandle;
     int instTextureCoordinatesHandle;
 
-    static const int instPositionsOffsetBytes = sizeof(GLfloat) * 0;
-    static const int instTextureCoordinatesOffsetBytes = sizeof(GLfloat) * 2;
-    static const int instScalesOffsetBytes = sizeof(GLfloat) * 6;
-    static const int instRotationsOffsetBytes = sizeof(GLfloat) * 8;
-    static const int instAlphasOffsetBytes = sizeof(GLfloat) * 9;
-    static const int instStretchInfoOffsetBytes = sizeof(GLfloat) * 10;
-    static const int instStretchXsAddOffsetBytes = sizeof(GLfloat) * 2;
-    static const int instStretchYsAddOffsetBytes = sizeof(GLfloat) * 6;
-    static const int instValuesSizeBytes = sizeof(GLfloat) * 20;
+    static const uintptr_t instPositionsOffsetBytes = sizeof(GLfloat) * 0;
+    static const uintptr_t instTextureCoordinatesOffsetBytes = sizeof(GLfloat) * 2;
+    static const uintptr_t instScalesOffsetBytes = sizeof(GLfloat) * 6;
+    static const uintptr_t instRotationsOffsetBytes = sizeof(GLfloat) * 8;
+    static const uintptr_t instAlphasOffsetBytes = sizeof(GLfloat) * 9;
+    static const uintptr_t instStretchInfoOffsetBytes = sizeof(GLfloat) * 10;
+    static const uintptr_t instStretchXsAddOffsetBytes = sizeof(GLfloat) * 2;
+    static const uintptr_t instStretchYsAddOffsetBytes = sizeof(GLfloat) * 6;
+    static const uintptr_t instValuesSizeBytes = sizeof(GLfloat) * 20;
 
 private:
     bool writeToDynamicInstanceDataBuffer(const ::SharedBytes &data, GLuint targetOffsetBytes);
